ScaffoldMessenger + refactor
