+++ lib/data/controllers/search.dart
@@ -90,7 +90,7 @@ class SearchController {
     return results;
   }
 
-  List<Searchable> getSearchables(BuildContext context, _scaffoldKey, callback) {
+  List<Searchable> getSearchables(BuildContext context, updateCallback) {
     List<Searchable> searchables = [];
 
     List<Message> messages = <List<Message>>[
@@ -105,7 +105,7 @@ class SearchController {
           child: MessageTile(message),
           onTap: () {
             Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => MessageView([message], _scaffoldKey, callback)));
+                builder: (context) => MessageView([message], updateCallback)));
           },
         ),  
     )));
