+++ lib/helpers/archivemessage.dart
@@ -3,14 +3,16 @@ import 'package:filcnaplo/data/context/app.dart';
 import 'package:filcnaplo/data/models/message.dart';
 import 'package:filcnaplo/generated/i18n.dart';
 
-Future archiveMessage(_scaffoldKey, BuildContext context, Message message, bool put, Function callback) async {
+Future archiveMessage(BuildContext context, Message message, bool archiving, Function updateCallback) async {
   int sentByUser = (message.sender == app.user.realName) ? 1 : 0;
-  int oldPlace = put ? sentByUser : 2;
-  int newPlace = put ? 2 : sentByUser;
+  int oldPlace = archiving ? sentByUser : 2;
+  int newPlace = archiving ? 2 : sentByUser;
 
-  app.user.kreta.trashMessage(put, message.id);
-  if(put && !(_scaffoldKey.currentState == null)) { //searchből nem kap rendes scaffoldkeyt
-    _scaffoldKey.currentState.showSnackBar(SnackBar(
+  app.user.kreta.trashMessage(archiving, message.id);
+  if (archiving) {
+    ScaffoldMessenger.of(context).hideCurrentSnackBar();
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
       content: Text(I18n
           .of(context)
           .messageDeleted),
@@ -25,7 +27,7 @@ Future archiveMessage(_scaffoldKey, BuildContext context, Message message, bool 
           app.user.sync.messages.data[oldPlace].add(message);
           app.user.sync.messages.data[oldPlace].sort((a, b) => a.date.compareTo(b.date));
           message.deleted = false;
-          callback();
+          updateCallback();
         },
       ),
     ));
@@ -33,6 +35,6 @@ Future archiveMessage(_scaffoldKey, BuildContext context, Message message, bool 
   app.user.sync.messages.data[oldPlace].removeWhere((msg) => msg.id == message.id);
   app.user.sync.messages.data[newPlace].add(message);
   app.user.sync.messages.data[newPlace].sort((a, b) => a.date.compareTo(b.date));
-  message.deleted=put;
-  callback();
+  message.deleted=archiving;
+  updateCallback();
 }
