+++ lib/ui/pages/frame.dart
@@ -84,7 +84,7 @@ class _PageFrameState extends State<PageFrame> {
 
     switch (app.selectedPage) {
       case 0:
-         pageContent = HomePage(_navItemSelected,_homeKey);
+         pageContent = HomePage(_navItemSelected);
         break;
       case 1:
         pageContent = EvaluationsPage(_homeKey);
@@ -99,7 +99,7 @@ class _PageFrameState extends State<PageFrame> {
         pageContent = AbsencesPage(_homeKey);
         break;
       default:
-        pageContent = HomePage(_navItemSelected,_homeKey);
+        pageContent = HomePage(_navItemSelected);
         break;
     }
